testing: de-duplicate code
