+++ infra/test/integration/suffix_example/suffix_example_test.go
@@ -12,142 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package suffix_buckets
+package simple_example
 
 import (
-	"io"
-	"net/http"
-	"strings"
 	"testing"
-	"time"
 
-	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/gcloud"
-	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/tft"
-	"github.com/GoogleCloudPlatform/cloud-foundation-toolkit/infra/blueprint-test/pkg/utils"
-	"github.com/gruntwork-io/terratest/modules/terraform"
-	"github.com/stretchr/testify/assert"
+	test "github.com/terraform-google-modules/dynamic-python-webapp/test/integration"
 )
 
 func TestSuffixExample(t *testing.T) {
-	example := tft.NewTFBlueprintTest(t)
-
-	example.DefineApply(func(assert *assert.Assertions) {
-		example.DefaultApply(assert)
-
-		// This module deploys a 'placeholder' Firebase Hosting release early
-		// in the process, to prevent a "Site Not Found" displaying when Terraform
-		// has finished applying, but the deployment is not yet complete.
-		//
-		// This extension of apply is meant to emulate that behavior. We confirm
-		// the placeholder behavior here to boost confidence that the frontend test in
-		// example.DefineVerify proves the placeholder page is replaced.
-		//
-		// If the check is flaky, remove it in favor of
-		// a simpler HTTP request.
-		//
-		// https://github.com/GoogleCloudPlatform/terraform-dynamic-python-webapp/issues/64
-		firebase_url := terraform.OutputRequired(t, example.GetTFOptions(), "firebase_url")
-		t.Log("Firebase Hosting should be running at ", firebase_url)
-		assertResponseContains(assert, firebase_url, "Your application is still deploying")
-	})
-
-	example.DefineVerify(func(assert *assert.Assertions) {
-		example.DefaultVerify(assert)
-
-		projectID := example.GetTFSetupStringOutput("project_id")
-		firebase_url := terraform.OutputRequired(t, example.GetTFOptions(), "firebase_url")
-		server_service_name := terraform.OutputRequired(t, example.GetTFOptions(), "server_service_name")
-		client_job_name := terraform.OutputRequired(t, example.GetTFOptions(), "client_job_name")
-
-		flagshipProduct := "Sparkly Avocado"
-		region := "us-central1"
-		t.Logf("Using Project ID %q", projectID)
-
-		{
-			// Check that the Cloud Storage API is enabled
-			services := gcloud.Run(t, "services list", gcloud.WithCommonArgs([]string{"--project", projectID, "--format", "json"})).Array()
-			match := utils.GetFirstMatchResult(t, services, "config.name", "storage.googleapis.com")
-			assert.Equal("ENABLED", match.Get("state").String(), "storage service should be enabled")
-		}
-
-		{
-			// Check that the expected Cloud Run service is deployed, is serving, and accepts unauthenticated requests
-			cloudRunServices := gcloud.Run(t, "run services list", gcloud.WithCommonArgs([]string{"--filter", "metadata.name=" + server_service_name, "--project", projectID, "--format", "json"})).Array()
-			nbServices := len(cloudRunServices)
-			assert.Equal(1, nbServices, "we expected a single Cloud Run service called %s to be deployed, found %d services", server_service_name, nbServices)
-			match := utils.GetFirstMatchResult(t, cloudRunServices, "kind", "Service")
-			serviceURL := match.Get("status.url").String()
-			assert.Truef(strings.HasSuffix(serviceURL, ".run.app"), "unexpected service URL %q", serviceURL)
-			t.Log("Cloud Run service is running at", serviceURL)
-
-			// The Cloud Run service is the app's API backend (it does not serve the Avocano homepage)
-			assertResponseContains(assert, serviceURL, "/api", "/admin")
-
-			// The data is populated by two Cloud Run jobs, so wait for the final job to finish before continuing.
-			// A job execution is completed if it has a completed time.
-			isJobFinished := func() (bool, error) {
-				clientJobExecs := gcloud.Run(t, "run jobs executions list ", gcloud.WithCommonArgs([]string{"--filter", "metadata.name~" + client_job_name, "--project", projectID, "--region", region, "--format", "json"})).Array()
-
-				if len(clientJobExecs) == 0 {
-					t.Log("Cloud Run job been executed. Retrying...")
-					return true, nil
-				}
-
-				match := utils.GetFirstMatchResult(t, clientJobExecs, "kind", "Execution")
-				completionTime := match.Get("status.completionTime").String()
-
-				if completionTime == "" {
-					// retry
-					t.Log("Cloud Run job execution hasn't completed. Retrying...")
-					return true, nil
-				}
-
-				t.Log("Cloud Run job completed", completionTime)
-				assert.NotEqual(completionTime, "", "completedTime must have a value")
-
-				succeededCount := match.Get("status.succeededCount").Int()
-				assert.Equal(succeededCount, int64(1), "succeededCount must not be 0")
-
-				return false, nil
-			}
-			utils.Poll(t, isJobFinished, 10, time.Second*10)
-
-			// The API must return a list that includes our flagship product
-			assertResponseContains(assert, serviceURL+"/api/products/", flagshipProduct)
-		}
-		{
-			// Check that the Avocano front page is deployed to Firebase Hosting, and serving
-			t.Log("Firebase Hosting should be running at ", firebase_url)
-			assertResponseContains(assert, firebase_url, "<title>Avocano</title>")
-		}
-	})
-	example.Test()
-}
-
-func assertResponseContains(assert *assert.Assertions, url string, text ...string) {
-	code, responseBody, err := httpGetRequest(url)
-	assert.Nil(err)
-	assert.GreaterOrEqual(code, 200)
-	assert.LessOrEqual(code, 299)
-	for _, fragment := range text {
-		assert.Containsf(responseBody, fragment, "couldn't find %q in response body", fragment)
-	}
-}
-
-func assertErrorResponseContains(assert *assert.Assertions, url string, wantCode int, text string) {
-	code, responseBody, err := httpGetRequest(url)
-	assert.Nil(err)
-	assert.Equal(code, wantCode)
-	assert.Containsf(responseBody, text, "couldn't find %q in response body", text)
-}
-
-func httpGetRequest(url string) (statusCode int, body string, err error) {
-	res, err := http.Get(url)
-	if err != nil {
-		return 0, "", err
-	}
-	defer res.Body.Close()
-
-	buffer, err := io.ReadAll(res.Body)
-	return res.StatusCode, string(buffer), err
+	test.AssertExample(t)
 }
